quotable layout
