+++ client2/src/pages/randomQuote/randomQuote.css
@@ -1,4 +1,4 @@
-.App {
+/* .App {
     display: flex;
     justify-content: center;
     min-height: 100vh;
@@ -47,4 +47,4 @@
     color: #FFF;
     box-shadow: 0 8px 16px 0 rgba(0,0,0,0.2), 0 6px 20px 0 rgba(0,0,0,0.19);
 }
-
+ */
